requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-cors-headers==3.10.1
djangorestframework==3.12.4
dnspython==2.2.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openpyxl==3.1.2
packaging==23.0
pymongo==4.3.3
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.2
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
